Move Pattern to Control fragment
