+++ app/src/main/java/rtandroid/ballsort/gui/fragments/ControlFragment.java
@@ -1,5 +1,7 @@
 package rtandroid.ballsort.gui.fragments;
 
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Handler;
 import android.support.v4.app.Fragment;
@@ -9,18 +11,22 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.GridLayout;
 import android.widget.Switch;
 import android.widget.TextView;
 
 import rtandroid.ballsort.MainActivity;
 import rtandroid.ballsort.R;
 import rtandroid.ballsort.blocks.AStateBlock;
+import rtandroid.ballsort.blocks.color.ColorType;
 import rtandroid.ballsort.blocks.loops.ResetLoop;
 import rtandroid.ballsort.blocks.loops.SortLoop;
 import rtandroid.ballsort.gui.ColorView;
 import rtandroid.ballsort.services.ResetService;
 import rtandroid.ballsort.services.SortService;
+import rtandroid.ballsort.settings.Constants;
 import rtandroid.ballsort.settings.DataState;
+import rtandroid.ballsort.settings.Settings;
 import rtandroid.ballsort.settings.SettingsManager;
 import rtandroid.ballsort.util.Utils;
 
@@ -46,6 +52,8 @@ public class ControlFragment extends Fragment
 
     private static Intent mSortIntent = null;
     private static Intent mResetIntent = null;
+
+    private ColorView mSelectedView = null;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
@@ -99,6 +107,57 @@ public class ControlFragment extends Fragment
         new ResetLoop();
 
         mUiUpdateHandler.postDelayed(mUiUpdateRunnable, REFRESH_RATE_MS);
+
+        GridLayout grid = (GridLayout) view.findViewById(R.id.patternGrid);
+        Settings settings = SettingsManager.getSettings();
+        grid.setColumnCount(Constants.PATTERN_COLUMNS_COUNT);
+        grid.setRowCount(Constants.PATTERN_COLUMNS_SIZE);
+        grid.removeAllViews();
+
+        for (int col = Constants.PATTERN_COLUMNS_COUNT-1; col>=0 ; col--)
+        {
+            for(int row = 0; row < Constants.PATTERN_COLUMNS_SIZE; row++)
+            {
+                View gview = inflater.inflate(R.layout.grid_view_entry, null, false);
+                ColorView cv = (ColorView) gview.findViewById(R.id.CV);
+                cv.setColor(settings.Pattern[col][row].getPaintColor());
+                int finalRow = Constants.PATTERN_COLUMNS_SIZE-row-1;
+                int finalCol = col;
+                cv.setOnClickListener(new View.OnClickListener()
+                {
+                    @Override
+                    public void onClick(View v)
+                    {
+                        mSelectedView = cv;
+
+                        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+                        builder.setTitle("Please choose a color:");
+                        builder.setItems(ColorType.names, new DialogInterface.OnClickListener()
+                        {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which)
+                            {
+                                Settings settings = SettingsManager.getSettings();
+                                ColorType type = ColorType.values()[which];
+                                settings.Pattern[finalCol][finalRow] = type;
+                                mSelectedView.setColor(type.getPaintColor());
+                                Log.d(MainActivity.TAG, "New color "+finalCol+"  "+finalRow+" is "+type.name());
+                            }
+                        });
+
+                        builder.setNegativeButton("Cancel", new DialogInterface.OnClickListener()
+                        {
+                            public void onClick(DialogInterface dialog, int id) {}
+                        });
+
+                        AlertDialog dialog = builder.create();
+                        dialog.show();
+                    }
+                });
+                grid.addView(gview);
+            }
+        }
+
 
         return view;
     }
